Move importing of mp to lib
